requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
pyyaml==6.0.1
joblib==1.4.0
shap==0.45.0
httpx==0.27.0
python-multipart==0.0.9
pytest==8.2.0
pytest-asyncio==0.23.6
